minor comments and body length is represented in unsigned long long int

+++ inc/http_response.h
@@ -19,6 +19,7 @@ struct http_response_head
 	// body stream
 };
 
+// returns 1 for success and 0 for failure
 // may fail only dure to memory allocation failure errors
 int init_http_response_head(http_response_head* hr_p);
 
@@ -40,7 +41,7 @@ void print_http_response_head(const http_response_head* hr_p);
 
 // returns 1 for success and 0 for failure
 // it is a util function, hence it does not return any error code to specify why it failed
-int init_http_response_head_from_http_request_head(http_response_head* hrp_p, const http_request_head* hrq_p, int status, size_t content_length_val);
+int init_http_response_head_from_http_request_head(http_response_head* hrp_p, const http_request_head* hrq_p, int status, unsigned long long int content_length_val);
 
 const dstring* find_acceptable_content_encoding_for_response(const http_request_head* hrq_p);
